Big rewrite of ctx_place.c
Hopefully the large rewrite to ctx_place should finally put to rest
issue #1: Missing Variants, where released-cortex-run-calls-one-by-one pipeline
was outperforming ninja-cortex is some test cases.  Tests suggest this is now
fixed, will run on staph test case to verify.

ctx_place.c: re-write and the follow changes:
* Added --match --mismatch --gapopen --gapextend alignment param args
* removed --ploidy and genotyping code outline -> handled elsewhere now
* removed --maxbr option (max branches) since doing pairwise alignment not MSA
* use different alignment scoring for flank search vs allele alignment
* add --out <out.vcf> option otherwise print to STDOUT
* more sensitive filter on 3' flank not matching: kmer/2 bp must MATCH

vcf_parsing.c:
* fixed BUG setting lf/rf tags
* improved adding new info tag -> look for empty slot instead of constantly
  increasing memory
* added vcf_info_tag_del()

+++ src/tools/ctx_supernodes.c
@@ -90,8 +90,8 @@ static inline void dot_print_edges2(hkey_t node, BinaryKmer bkmer, Edges edges,
 
     if(node < next_nodes[i] || (node == next_nodes[i] && side0 <= side1)) {
       fprintf(fout, "  node%zu:%c -> node%zu:%c [color=\"blue\"]\n",
-              snode0.nodeid, coords[side0],
-              snode1.nodeid, coords[side1]);
+              (size_t)snode0.nodeid, coords[side0],
+              (size_t)snode1.nodeid, coords[side1]);
     }
   }
 }
